Extract common functionality to base class
