+++ Assets/Editor/BasePreferenceProvider.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 16af8b0008d004032aeca1150da4a82f
+timeCreated: 1469108734
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
